+++ src/object/values.js
@@ -1,5 +1,6 @@
 import _reduce from '../_internals/_reduce.js'
 import _appendǃ from '../_internals/_appendǃ.js'
+import keys from './keys.js'
 
 /**
  * @name values
@@ -16,6 +17,6 @@ import _appendǃ from '../_internals/_appendǃ.js'
  * values({ a: 1, b: 2, c: 3 }) // => [1, 2, 3]
  */
 const values = obj =>
-  _reduce((k, acc) => _appendǃ(acc, obj[k]), [], Object.keys(obj))
+  _reduce((k, acc) => _appendǃ(acc, obj[k]), [], keys(obj))
 
 export default values