+++ CHANGELOG.md
@@ -16,7 +16,12 @@
 ### Improved
 
 - `values` now uses the built in `keys` function instead of `Object.keys`
+- `draft` now uses the built in `keys` function instead of `Object.keys`
 - Greatly expanded the unit testing around `values`
+- Expanded the unit tests for `groupBy`
+- Expanded the unit tests for `gt` and `gte`
+- Expanded the unit tests for `difference`
+- Added an extra unit test for `plan`
 
 ### Fixed